make mypy run clean
Finish making mypy run with clean output: from 532 errors down to 0, by ignoring problems with imports, adding some type declarations (some loose ones), and excluding profile_polymerize.

This makes it useful in CI and it _should_ help catch problems with Python 3 conversion steps like the iterators and str vs. bytes.

+++ runscripts/reconstruction/fold_changes/calculate_fc.py
@@ -72,8 +72,8 @@ def load_src(attempt_match):
 			if attempt_match:
 				fcs = np.abs(fcs)  # Bad!! - ignores annotated condition comparison regulation direction
 			tf_data[gene] = {
-				'mean': np.mean(fcs),
-				'std': np.std(fcs, ddof=1),
+				'mean': float(np.mean(fcs)),
+				'std': float(np.std(fcs, ddof=1)),
 				}
 		processed_data[tf] = tf_data
